Fix command not found error on macOS
package.json inside dist directory is the cause of this error.
It may be related to unix or nvm environments. So, TypeScript
copies package.json to satisfy the import in sendToken.ts. To
get rid of this file, I removed api field in package.json and
add it to constants.